app router: Fix infinite redirect loop in MPA navigation
Not 100% convinced that this is the correct fix but it's the best I can think of.

Previously, we would sometimes call location.assign()/.replace() hundreds of times (or more) as I described in https://github.com/vercel/next.js/issues/48438 and https://github.com/vercel/next.js/issues/48309#issuecomment-1512290958. Sometimes this would just make things slow but the navigation would eventually succeed; sometimes this would just hang the browser.

Now we trigger it only once (or—for a reason I don't quite understand—twice in dev, as you can see in the test) and never commit that render.

This also fixes the bug I mentioned in https://github.com/vercel/next.js/issues/48438#issuecomment-1528649776 where usePathname() and useSearchParams() would return the page we are navigating to (even if that's an external page wholly unrelated to our site!).

Fixes #48309, fixes #48438.