Added public/static and root path to matcher
Homepage isn't covered by the suggested matcher config which is a common case and a desired behavior. Also a `public/static` folder that comes with Next.js by default is also not covered in this example. Absence of these values produces a [lot of confusion](https://github.com/vercel/next.js/discussions/38615) for newbies.