Rust daemon v2 (vercel/turbo#4791)
### Description

This reintroduces the rust daemon with some additional patches to help
with some critical shortcomings of the initial canary.

In short:
- kqueue performance is pretty awful and would gladly pin a core on your
CPU for multiple minutes while registering watchers. this is an opt-in
feature on macOS so we dropped it in favour of fsevents
- we didn't have timeouts for daemon interactions, meaning it would hang
while waiting for the daemon to process a request. the timeout is now
set to 1s
- we had a few outstanding bugs

With the above addressed, we are going to try another canary.

---------

Co-authored-by: Greg Soltis <greg.soltis@vercel.com>