Apply pending transition to context when computing resolve options (vercel/turbo#4901)
### Description

This fixes an issue where references to assets imported with transition
annotations would not be resolved according to the proper resolve
options context of the transition, but that of the previous asset
context instead.

This only fixes parts of the underlying issue. A more complete solution
is propoed in WEB-1042.

link WEB-1033