Re-throw module.error when instantiating runtime modules (vercel/turbo#4918)
### Description

Runtime modules can be executed multiple times within the same context.
However, when such a module encounters a runtime error during execution,
this error is cached in module.error and not properly re-thrown when the
module is required again from `getOrInstantiateRuntimeModule`, which
means that on later runs we’ll silently fail to instantiate the module,
and run into any errors that expects a global side-effect from the
module.

link WEB-1045