Create _document.tsx
To prevent the flash of unstyled content in my Next.js application, I added a custom Document component. In this component, I included a link to the Ant Design CSS stylesheet within the <Head> section. This ensures that the styles from Ant Design are loaded and applied to the page during server-side rendering (SSR). By extending the default Document component, I was able to customize the HTML structure of the page and include the necessary styles, providing a seamless and styled user experience