Update prerender-error.md
During the function development some time ago, we found that there was a similar error report. It said that there was another abnormal pre-rendered page route. After a series of investigations according to the document, we found that the modules under the routing folder referenced a large number of next. replace and push, this may be an error caused when the current route route has not been initialized, we can solve this problem by adding isReady judgment at the root route of the folder, adding this information may also better help other developers to troubleshoot Problems with pre-rendering.