Prepare RouterContentSource for basePath (vercel/turbo#5218)
### Description

This is part 2 of [WEB-993](https://linear.app/vercel/issue/WEB-993)
basePath support. A few of our next-specific content sources will need
to be scoped under the `basePath` (like `_next/image` and
`__nextjs_original-stack-frame`). These are currently served with a
`RouterContentSource`, but it didn't have support for arbitrary
prefixes.

We _could_ have changed the subpath for these sources to include the
`basePath`, but that would require reading the `next_config.base_path()`
in the
[source](https://github.com/vercel/next.js/blob/2b1f0d9351610b04d01638efed19252ca81d0023/packages/next-swc/crates/next-dev/src/lib.rs#L413-L423)
method, and it would invalidate our entire call graph whenever the
`next.config.js` changed. Not a good choice.


### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->