ci(workflow): loosen condtions for uploading turbopack bytesize (#50820)
### What?

It turned out the recommendation https://github.com/vercel/next.js/pull/50600/files#r1212902688 is overly strict, even if it's valid build work we skips to upload data and only triggers for the release.  This is an example workflow https://github.com/vercel/next.js/actions/runs/5182537384/jobs/9339461369 that build ran but upload skipped.

We are fine with same bytesize numbers with cache hit actually, and can filter it out in worst cases - so loosening conditions for now.