Update 10-router-handlers.mdx
When I first saw this section, I got confused due to the wrongly named file ending and the multiline TypeScript syntax.

What I did was added a TS and JS switcher. I also changed the TypeScript syntax for { params } : { params: { slug: string }}

This is how I saw it written in the example in 05-dynamic-routes of the documentation and I think it is much easier to read and less confusing.