Update 01-images.mdx (#51532)
Due to the size of grid tracks specified in the CSS I amended the usage of the HTML `sizes` attribute.

The media can be ascertained from `gridTemplateColumns` and `gridGap` values, this change results in smaller images also being available in the srcset for use with sizes, which should be more performant.