Add default modularizeImports that breaks up lucide-react, mui/icons-material, lodash, react-bootstrap (#50900)
## What?

Adds a default modularizeImports that ensures `lucide-react`, `@mui/icons-material`, `lodash`, and `react-bootstrap` do not cause a massive amount of modules to be compiled in development.
For example `@mui/icons-material` re-exports 10600 (10K+) modules from a single file. Even though it does not affect production bundle size it does affect build time, especially in development, and even more so when your filesystem is slow (e.g. what we're seeing on Windows). 

Related issue: #48748 (not closing)




Co-authored-by: Shu Ding <3676859+shuding@users.noreply.github.com>